[ci] Add Travis Setup.
This doesn't check that the output of the script is correct but it
should do the job. Fixes: #46

We have done quite a few tweaks to the script, in particular:

- we suffered https://github.com/ocaml/opam/issues/2767
- the script is much less verbose now [`$BENCH_DEBUG` can be use to
  increase verbosity.
- `perf` doesn't work in Travis `<unsupported>`, this was not fun.

It would be great to upload artifacts but I'm afraid Travis doesn't
let us do this (yet?) Thus, we prefer to keep `stdout` unredirected on
a few places.

We can much improve things, including being able to use the cache, but
this will be done in future PRs.