Don't require config to be topologically sorted
Instead we perform a stable sort using the dependencies generated by
paring the output of `opam install --show-actions`.  Note that we depend
on the output actions matching
`∗\s*install\s*<package-name-without-spaces>`.  Note the special unicode
star character.  Perhaps we should only look for `^\s*.\s*intstall\s*`
instead, in case opam changes what character they use to indicate an
install action?  Or `\sinstall\s*`, in case the character goes away
entirely?  (But then we have a bit of trouble if we ever try to install
a package called "install".)

We do a manual sort because we want to preserve user-ordering insofar as
it is possible.  In particular, we want to allow bench initiators to
declare that certain packages should be installed as late as possible,
by putting them at the end of the list, so that if they interrupt the
bench early, they get the packages they care most about.  (I could not
figure out how to get `opam install --show-actions` to be a stable
sort.)

Closes #3